features!
You can now scan for devices on the network to control. Use the libnetat or libnetat.py interfacename scan command, which will return the MAC addresses of available devices. To control a device non-interactively, pass the interfacename and the macaddress as parameters.

In interactive mode, you can type scan to scan for devices. You can manually set a MAC address of a device to send commands to by using the setmac command. To exit interactive mode gracefully, type exit.